Pass puppet agent's originally configured env to node terminus
The loop in configurer.rb to determine an agent's environment has a bug where
the originally configured environment is forgotten if an ENC supplies a new
environment. When using an ENC that only sometimes overrides an agent's
environment, this can lead to a situation where on the first check the ENC
overrides an environment, but on the second it classifies the node such that
it can choose its own environment, and due to this bug the node will choose
the environment returned by the first call to the ENC rather than its
configured environment.

The solution here is to store the originally configured environment and pass
it through to the node terminus. This commit does not update the terminus
itself to use this new information.