(maint) Optimize tags
There are a bunch of places where tags are merged from one object into
another and we already know they are valid, so it's unnecessary to
recheck them. Some code was already in place to reduce rechecking. This
commit applies that change in more places.

Also, it appears unnecessary to duplicate a tagset every time it is
returned. We do this a lot so removing the dup call reduces garbage
significantly.