(maint) Refactor configurer_spec.rb to use lets
This is the first step in a series of commits to refactor
configurer_spec.rb to not be test order dependent.

In this step, use of @agent, @catalog and @facts instance variables are
changed to instead use rspec let() operations.

This revealed several order issues where tests assumed that @catalog was
set a specific way.