(PUP-8657) Resolve deferred Sensitive with unwrapped and set flag
Before this, if resolution of a deferred value resulted in a Sensitive
and the resource attribute it is assigned to was not already marked as
sensitive the sensitive-flag on that attribute would be lost.

Now, the resolution of a deferred Sensitive value works as if had been
set as a direct Sensitive value and the `sensitive_parameters` is
updated in the resolution process.