(PUP-8733) Store full file path in LOADED_FEATURES in autoloader
The implementations of `require` in JRuby and MRI don't check the same
paths in `$LOADED_FEATURES`. Both check a version of the relative path,
but MRI checks with the file extension and JRuby checks without. Both
check the absolute path, so if we save that instead we should get
consistent behavior across platforms.