(PUP-8971) Make `decrypt` always return `Sensitive`
The typical use case is that a decrypted value needs to be handled such
as the secret is not revealed by accident. Therefore, the returned value
from `decrypt` is now always a `Sensitive` value as that prevents it
from being revealed in reports and logs.

This also updates the documentation of `decrypt`